requirements.txt
altgraph==0.17.4
certifi==2025.8.3
charset-normalizer==3.4.3
customtkinter==5.2.2
darkdetect==0.8.0
dotenv==0.9.9
idna==3.10
mutagen==1.47.0
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
python-dotenv==1.1.1
pywin32-ctypes==0.2.3
redis==6.4.0
requests==2.32.5
setuptools==80.9.0
spotipy==2.25.1
urllib3==2.5.0
yt-dlp==2025.9.5
